+++ packages/core/src/scripts/cli-index.ts
@@ -1,7 +1,7 @@
-import { cli } from './cli';
+import { cliMeow } from './cli';
 import { ExitError } from './utils';
 
-cli(process.cwd(), process.argv.slice(2)).catch(err => {
+cliMeow(process.cwd(), process.argv.slice(2)).catch(err => {
     if (err instanceof ExitError) {
         process.exit(err.code);
     }