+++ packages/ali-sms-plugin/CHANGELOG.md
@@ -1,5 +1,141 @@
 # @picker-cc/ali-sms-plugin
 
+## 22.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@22.0.0
+
+## 21.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@21.0.0
+
+## 20.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@20.0.0
+
+## 19.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@19.0.0
+
+## 18.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@18.0.0
+
+## 17.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@17.0.0
+
+## 16.0.0
+
+### Major Changes
+
+-   y
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+    -   @picker-cc/common@16.0.0
+
+## 15.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@15.0.0
+
+## 14.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@14.0.0
+
+## 13.0.0
+
+### Major Changes
+
+-   Y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/common@13.0.0
+
+## 12.0.0
+
+### Major Changes
+
+-   Y
+-   Y
+-   Y
+
+### Patch Changes
+
+-   @picker-cc/common@10.0.1
+
+## 12.0.0-alpha.1
+
+### Major Changes
+
+-   alpha
+
+## 12.0.0-alpha.0
+
+### Major Changes
+
+-   alpha
+
 ## 10.0.0
 
 ### Major Changes