+++ packages/custom-auth/CHANGELOG.md
@@ -1,5 +1,150 @@
 # auth
 
+## 21.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@22.0.0
+
+## 20.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@21.0.0
+
+## 19.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@20.0.0
+
+## 18.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@19.0.0
+
+## 17.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@18.0.0
+
+## 16.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@17.0.0
+
+## 15.0.0
+
+### Major Changes
+
+-   y
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+    -   @picker-cc/core@16.0.0
+
+## 14.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@15.0.0
+
+## 13.0.0
+
+### Major Changes
+
+-   y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@14.0.0
+
+## 12.0.0
+
+### Major Changes
+
+-   Y
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@13.0.0
+
+## 11.0.0
+
+### Major Changes
+
+-   Y
+-   Y
+-   Y
+
+### Patch Changes
+
+-   Updated dependencies
+-   Updated dependencies
+-   Updated dependencies
+    -   @picker-cc/core@12.0.0
+
+## 11.0.0-alpha.1
+
+### Major Changes
+
+-   alpha
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@12.0.0-alpha.1
+
+## 11.0.0-alpha.0
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @picker-cc/core@12.0.0-alpha.0
+
 ## 10.0.0
 
 ### Major Changes