+++ packages/wechat-plugin/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @picker-cc/wechat-plugin
 
+## 22.0.0
+
+### Major Changes
+
+-   y
+
+## 21.0.0
+
+### Major Changes
+
+-   y
+
+## 20.0.0
+
+### Major Changes
+
+-   y
+
+## 19.0.0
+
+### Major Changes
+
+-   y
+
+## 18.0.0
+
+### Major Changes
+
+-   y
+
+## 17.0.0
+
+### Major Changes
+
+-   y
+
+## 16.0.0
+
+### Major Changes
+
+-   y
+-   y
+
+## 15.0.0
+
+### Major Changes
+
+-   y
+
+## 14.0.0
+
+### Major Changes
+
+-   y
+
+## 13.0.0
+
+### Major Changes
+
+-   Y
+
+## 12.0.0
+
+### Major Changes
+
+-   Y
+-   Y
+-   Y
+
+## 12.0.0-alpha.1
+
+### Major Changes
+
+-   alpha
+
+## 12.0.0-alpha.0
+
+### Major Changes
+
+-   alpha
+
 ## 11.0.0
 
 ### Major Changes